Renamed task1 to document_processor
